+++ swift/Sources/LibSignalClient/SealedSender.swift
@@ -173,9 +173,9 @@ public func sealedSenderMultiRecipientEncrypt(_ content: UnidentifiedSenderMessa
         let sessionHandles = sessions.map { $0.unsafeNativeHandle }
         return try content.withNativeHandle { contentHandle in
             return try recipientHandles.withUnsafeBufferPointer { recipientHandles in
-                let recipientHandlesBuffer = SignalBorrowedSliceOfProtocolAddress(base: recipientHandles.baseAddress, length: UInt(recipientHandles.count))
+                let recipientHandlesBuffer = SignalBorrowedSliceOfProtocolAddress(base: recipientHandles.baseAddress, length: recipientHandles.count)
                 return try sessionHandles.withUnsafeBufferPointer { sessionHandles in
-                    let sessionHandlesBuffer = SignalBorrowedSliceOfSessionRecord(base: sessionHandles.baseAddress, length: UInt(sessionHandles.count))
+                    let sessionHandlesBuffer = SignalBorrowedSliceOfSessionRecord(base: sessionHandles.baseAddress, length: sessionHandles.count)
                     return try withIdentityKeyStore(identityStore, context) { ffiIdentityStore in
                         try invokeFnReturningArray {
                             signal_sealed_sender_multi_recipient_encrypt($0,
